requirements.txt
Flask
requests
python-dotenv
websocket-client